refactor update module dependencies
- Replaced the `core.data` dependency with `core.domain`.
