+++ feature/employee/build.gradle.kts
@@ -24,7 +24,7 @@ android {
 }
 
 dependencies {
-    implementation(projects.core.data)
+    implementation(projects.core.domain)
 
     testImplementation(libs.hilt.android.testing)
     testImplementation(projects.core.testing)
