+++ feature/inventory/build.gradle.kts
@@ -24,7 +24,6 @@ android {
 }
 
 dependencies {
-    implementation(projects.core.data)
 
     testImplementation(libs.hilt.android.testing)
     testImplementation(projects.core.testing)
