+++ feature/item/src/main/java/com/casecode/pos/feature/item/ItemsViewModel.kt
@@ -19,7 +19,7 @@ import android.graphics.Bitmap
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.casecode.pos.core.data.utils.NetworkMonitor
+import com.casecode.pos.core.domain.utils.NetworkMonitor
 import com.casecode.pos.core.designsystem.component.SearchWidgetState
 import com.casecode.pos.core.domain.usecase.AddItemUseCase
 import com.casecode.pos.core.domain.usecase.DeleteItemUseCase
