+++ feature/reports/src/main/java/com/casecode/pos/feature/statistics/ReportsViewModel.kt
@@ -17,7 +17,7 @@ package com.casecode.pos.feature.statistics
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.casecode.pos.core.data.utils.NetworkMonitor
+import com.casecode.pos.core.domain.utils.NetworkMonitor
 import com.casecode.pos.core.domain.usecase.GetTodayInvoicesUseCase
 import com.casecode.pos.core.domain.utils.Resource
 import dagger.hilt.android.lifecycle.HiltViewModel
