+++ feature/supplier/src/main/kotlin/com/casecode/pos/feature/supplier/SupplierViewModel.kt
@@ -18,7 +18,7 @@ package com.casecode.pos.feature.supplier
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.casecode.pos.core.data.utils.NetworkMonitor
+import com.casecode.pos.core.domain.utils.NetworkMonitor
 import com.casecode.pos.core.designsystem.component.SearchWidgetState
 import com.casecode.pos.core.domain.usecase.AddSupplierUseCase
 import com.casecode.pos.core.domain.usecase.DeleteSupplierUseCase
